move diagnostic_utils to utils; refactoring
